Add support for multiuser to VnInjectProcess
VnInjectProcess did not have an option to check for session ID in processes, meaning that it would just inject into the first process with the desired name. This adds an option (bool) to enable injecting only to the currently active session ID. I haven't done any thorough testing, but I will be using this in my own project as well, so if any problems arise from the use of this option, I will be quick to patch them. Setting the new option to false should leave everything else untouched.